Make SO3 feature complete (for the mvp design) (#37)
## Description of change
This change adds the last planned methods for the feature complete SO3 class
as laid out in the design doc linked to the following task:
[Asana task](https://app.asana.com/0/1202178773526279/1202227479029247/f)

Also it adds a number of tests to cover those new methods. Please note that
this does not complete all tests planned for the SO3 class. There are some 
additional 'sanity check' tests to follow.

## Asana task link
[Asana task](https://app.asana.com/0/0/1202264433314345/f)
## Guide to reproduce test results.
```
bazel test //transforms:so3_test
```
## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [x] This change is covered by tests that are already landed, or in this PR.