Two-jet class (#86)
## Description of change
The planned time parameterized curve requires a class to describe a point on 
the curve which consists of a LieGroup transform and its first and second time 
derivatives in left tangent space. Such a class can be reused in other
applications - such as describing the state of a rigid actor in the number of 
degrees of freedom of the underlying LieGroup.
## Asana task link
[Asana task](https://app.asana.com/0/0/1202672342571864/f)
## Guide to reproduce test results.
```
bazel test //curves:two_jet_test
```
## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [x] This change is covered by tests that are already landed, or in this PR.