Fix function covarage regression in MVG tests. (#158)
## Description of change
Using a hard-coded dimensionality for Gaussian of '4' in the death tests here
created compiler generated functions that were unused. Strictly all that was
required to get full coverage was to re-use one of the contexpr dimensionality
values used in the other tests. However this creates a potential trap for 
future engineers using this, so we have gone for a more complete solution of 
type-parameterized tests for all three dimensionality values. One downside of
this is that this does make the sampling test quite heavyweight and now the
test suite takes 30s to run.

# Asana task link
[Asana task](https://app.asana.com/0/0/1203504667762793/f)
## Guide to reproduce test results.
```
bazel test //resim_core/math:multivariate_gaussian_test
./resim_core/scripts/run_codecov_locally.sh
```
## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [x] This change is covered by tests that are already landed, or in this PR.