Remove dimensionailty templating from Gaussian, making it fully dynamic (#194)
## Description of change
I regret the decision to template Gaussian to its dimensionality. While the 
compile time checking is noble, the requirement to know the dimensionality at
compile time is restrictive. In particular when modelling trajectories as a 
Gaussian, we need to know the number of control points, which seems 
inconvenient.

This change make Gaussian fully dynamic. It also adds additional unit tests
to compensate for the loss of the compile-time checking.
## Asana task link
[Asana task](https://app.asana.com/0/1202178773526279/1203716925502632/f)
## Guide to reproduce test results.
```
bazel test //resim-core/math:multivariate_gaussian_test
```
## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [x] This change is covered by tests that are already landed, or in this PR.