FramedGroup is_approx pays attention to frames (#332)
## Description of change
This PR contains two small but substantive design changes. I detail explicitly 
the two changes below. The rest of the 10 - or so - files touched in this PR are 
changes to tests that made assumptions that have been broken by these changes.
### Change One - is_approx pays attention to frames for framed groups.
```
  // Approximate comparison with another FramedGroup. This expects the frames
  // to match.
  bool is_approx(const FramedGroup<Group> &other) const;

  // Approximate comparison with an unframed Group. The frames are ignored.
  bool is_approx(const Group &other) const;
```
The above two methods are added to `FramedGroup` when comparing two framed 
groups it is expected that the frames will match. Comparison with 
unframed groups ignores the frames. Also a `.group()` getter is added so a user 
that want to only compare transforms of framed groups can do e.g.
```
framed_a.is_approx(framed_b.group());
```
Previously FramedGroup inherited the `is_approx` method from `Group` so many 
tests used this without care for the frames, hence the large number of changes.
### Change Two - FramedGroup `.interp()` defaults to re-use the from frame
We  recently changed `TCurve` and `DCurve` to use a consistent point frame 
instead of insisting that all point samples be a new frame. Given this new 
pattern it makes sense that `FramedGroup::interp` do something similar. It used 
to default to a new frame, but now it defaults to re-use the from frame of the 
group:
```
template <LieGroupType Group>
FramedGroup<Group> FramedGroup<Group>::interp(const double fraction) const {
  return interp(fraction, from());
};
```
The user can sill use the other overload of `interp` to set a different frame if 
they wish. This just changes the default behavior.

The reason this change is bundled in this PR and not in a follow-on is that this 
update makes the test repair required after change one significantly simpler.
## Asana task link
[Asana task](https://app.asana.com/0/1203294986954613/1204295910830761/f)
## Guide to reproduce test results.
```
bazel test //resim-core/transforms/...
bazel test //resim-core/curves/...
```
## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [x] This change is covered by tests that are already landed, or in this PR.