Fixes a problem with JSON packing (#372)
## Description of change

Fixes a bug with packing Frame type (enum=0) to JSON

The value zero is assumed to be some sort of none types, so the 
string doesn't get packed. This change introduces an explicit None
and bumps the numbers.

Also fixes the Frame test to actually test the metadata.

## Asana task link
[Asana task](https://app.asana.com/0/1204081589980528/1204409832423567/f)
## Guide to reproduce test results.
```
bazel test ...
```
## Checklist:
- [X] I have self-reviewed this change.
- [X] I have tested this change.
- [X] This change is covered by tests that are already landed, or in this PR.