Refactor core sim logger (#423)
## Description of change

Refactors logging system, to have a `logger` attached to all simulation units, as opposed to a single `LoggerUnit`. This enables things like "easy per-actor metrics", which are also implemented in this PR.

Not yet implemented: thread-safety on logger. I had a shot at this, but decided it's probably best to leave it for another PR, as I've already let some scope creep happen in this big PR... I *think* the right way to go about this is probably to attach `log_proto` and `add_channel` methods to the base simulation unit, and attach a mutex there, but up for debate.

## Asana task link

[Asana task](https://app.asana.com/0/home/1204275002848728/1204654202364071)

## Guide to reproduce test results.

`bazel test //resim_core/...`

## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [x] This change is covered by tests that are already landed, or in this PR.