Fix a bug where DCurve always picks the first segment (#458)
## Description of change
Currently, DCurve always picks the first segment. This was not noticed
in tests because we always used a circle in our unit tests and view
tests, so the extrapolated first segment still followed the course of
the DCurve coincidentally. This change fixes the issue and adds a zig
zag test which uses a DCurve geometry such that the old code does not
pass, but the fixed code does.

## Asana task link
[Asana
task](https://app.asana.com/0/1204498029712344/1204775059650430/f)
## Guide to reproduce test results.
```bash
full_test () {
   TARGET="$1"
   VALGRIND="valgrind --leak-check=full --show-leak-kinds=all --track-origins=yes --error-exitcode=1"

   bazel test --runs_per_test=100 "${TARGET}" && \
   bazel test --test_timeout=3600 --run_under="${VALGRIND}" "${TARGET}"
}
full_test //resim_core/curves:d_curve_test
```

## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [x] This change is covered by tests that are already landed, or in
this PR.