Remove scope request from device code flow. (#506)
## Description of change

Removes the scope request from the device code flow. We don't actually
need this,
we use permissions for API access now. It's causing auth0 to return a
"scope"
field in the token response, which the proto we're deserializing into
doesn't have.

It's unclear whether auth0 suddenly started returning this field or
whether something
we did caused them to start doing it (I suspect the latter, but I can't
figure out what).

## Asana task link
[Asana task](<!-- Enter task link here -->)
## Guide to reproduce test results.

    bazel run //resim/visualization/examples:example_00_rigid_transform
    
## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [x] This change is covered by tests that are already landed, or in
this PR.