"Devcontainer scripts allow for users renaming their base dir" (#3)
## Description of change
This PR assumes that users cloning open-core, will fairly regularly
rename the directory to
something more meaningful to them. In this case it would be nice if the
change were
reflected in the devcontainer workflow.

There is a slight fragility to this implementation in that the user
could call `.devcontainer/*`
from another directory e.g. `../devcontainer/*`, but we are already
making use of `$(pwd)`
so this at least does not make the situation worse.

## Guide to reproduce test results.
1. Rename `open-core` e.g. `mv open-core test-core`
2. Run `.devcontainer/run.sh` and verify that the working directory is
`test_core`
3. Repeat for other affected scripts
## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [ ] This change is covered by tests that are already landed, or in
this PR.

Co-authored-by: Ubuntu <ubuntu@ip-172-31-18-171.us-west-1.compute.internal>