Add an add_task() overload to ExecutorBuilder which supports multi fan-in (#82)
## Description of change
It's commonly required for a particular task to depend on multiple
preceding tasks to complete before it can run. This is currently quite
annoying to accomplish, so this change adds an overload for
ExecutorBuilder::add_task() which takes care of the nitty-gritty details
to make this work. All data returned by the dependent tasks are
currently passed as vectors to the task which depends on multiple
dependencies.

## Guide to reproduce test results.
```bash
bazel test //resim/simulator:executor_builder_test
```
## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [x] This change is covered by tests that are already landed, or in
this PR.