Add static registry for our dependencies (#132)
## Description of change
Not all of our dependencies (e.g.
[httplib](https://github.com/yhirose/cpp-httplib) and
[rules_ros2](https://github.com/mvukov/rules_ros2)) support bzlmod out
of the box at the moment so we cannot rely on the bazel central registry
for them. In anticipation of moving to bzlmod we'll set up our own
registry that we'll maintain for these dependencies. When we move to
bzlmod, we'll add the following lines to our `.bazelrc` to rely on this
registry first, and the central registry second:

```
common --registry=https://raw.githubusercontent.com/resim-ai/open-core/prod/main/bazel_registry/
common --registry=https://bcr.bazel.build
```

Hopefully, we'll be able to remove some of the entries in our registry
as they become supported in the central registry.

Note that we haven't likely added all of the things that rules_ros2
actually depends on, but we've at least satisfied the dependencies that
we actually use in our code.

## Guide to reproduce test results.
Build using this registry on a local branch.

## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [ ] This change is covered by tests that are already landed, or in
this PR.