Monolithic python package (#145)
## Description of change
Make a single python package for all non-ROS2 related python stuff in
open-core. Set up this package to automatically get its version from the
input to the draft release workflow. Prevent old versions from appearing
in the assets field of the drafted release.

## Guide to reproduce test results.
Run the draft release workflow on this branch with input "v0.0.0"

![image](https://github.com/resim-ai/open-core/assets/17681777/8be375da-1b95-4466-adb3-ef26b6726a5e)

Running this quickly again with "v99.99.99" since 0.0.0 is the default:
https://github.com/resim-ai/open-core/actions/runs/9554416021


![image](https://github.com/resim-ai/open-core/assets/17681777/489ac049-ea77-4f10-a1b7-3f1a89dcaff3)


## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [x] This change is covered by tests that are already landed, or in
this PR.