Make pod parser mutable (#177)
## Description of change
Currently the macro used to define a "Plain-Old-Data" (POD) getter for
POD structs doesn't allow mutable access to the contents of the struct
even if the struct is mutable. This is due to the way in which
`decltype()` functions. Particularly "unparenthesized class member
access expression[s]" cause decltype to get the declared value of the
field being accessed which is generally not a reference. We use
parentheses to ameliorate this. See
[here](https://en.cppreference.com/w/cpp/language/decltype) for more
details. We also added some tests to confirm this now works.

## Guide to reproduce test results.
```bash
bazel test //resim/converter:parser_test
```
## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [x] This change is covered by tests that are already landed, or in
this PR.