Add helper to kickoff a report based off of a test suite name (#184)
## Description of change

Add a helper function which allows us to kickoff a report easily based
on a batch id and a test suite allowlist. The idea is we can use this to
automatically kick of reports as the last step of batch metrics.

Also add mocks for the API client so we don't have to hit the API to
test this. The expectation is that this mocks library will grow as more
endpoints require mocking.

## Guide to reproduce test results.

```bash
bazel test //resim/orchestration/...
```

## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [x] This change is covered by tests that are already landed, or in
this PR.