Auto-generate go bindings (#165)
## Description of change
Modify the release workflow to:
 - Check that all needed tags are available before releasing.
- Create a new tag corresponding to each release with the format
vX.Y.Z-go which contains a single commit on top of vX.Y.Z. This commit
generates the go bindings and places them in `uuid_proto/` and
`metrics_proto/`. Note that this tag is only a tag and no special
release is created for go.

## Guide to reproduce test results.
Run the Draft Release action and make sure the go tag is properly
created. This one fails, but it fails after the go content since the
wheels can't be generated with the nonsensical test version I used for
it. It would pass if a real version were used:


https://github.com/resim-ai/open-core/actions/runs/10289413350/job/28477102187

## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [x] This change is covered by tests that are already landed, or in
this PR.