Introduce Metrics Status Override (#199)
## Description of change

This PR proposes to extend the ReSim Metrics Writer to enable the
passing through of a MetricsStatus that would override the default
calculation for how an overall test status is calculated. It enables
users flexibility to determine exactly how a test would fail based on
the metrics it contains. For example, warnings could be ignored, rather
than cause the whole test to be considered in a warning status.

## Guide to reproduce test results.

```
bazel test //...
```

## Checklist:
- [X] I have self-reviewed this change.
- [X] I have tested this change.
- [X] This change is covered by tests that are already landed, or in
this PR.