Handle failing tests due to bazel update (#253)
## Description of change
The bazel update causes a failure for tests that make use of the
`TEST_UNDECLARED_OUTPUTS_DIR` directory because `bazel` now employs the
`file` and `zip` commands which our devcontainer does not have. This
adds those to the devcontainer and temporarily keeps those tests from
using that directory so we can pass tests with the old devcontainer.
This has no affect on the tests' actual coverage since that directory
was only being used to save visualization logs for human inspection.

## Guide to reproduce test results.
CI checks.

## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [x] This change is covered by tests that are already landed, or in
this PR.

---------

Co-authored-by: Matt C <matt@resim.ai>