Deprecating Rechart in metrics (#248)
## Description of change
We are moving our charting library to Plotly, and are deprecating
metrics that used Rechart. This PR does a couple of things:

- [x] Groups deprecated metrics into a section denoting the metrics are
deprecated
- [x] Adds a note at the top of the section saying they are deprecated
(possibly add it to each chart individually to make it very obvious?)

## Guide to reproduce test results.

To test you can pull it locally and run it and see if things look good. 

## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [x] This change is covered by tests that are already landed, or in
this PR.