Fix proto versioning in wheels (#268)
## Description of change
Update the protobuf version so that our dependencies don't pick it to be
a release candidate (see
[here](https://registry.bazel.build/modules/rules_python/1.5.1)).

Also stop using deprecated rules for proto generation to remove some
warnings.

## Guide to reproduce test results.
Unit tests check that the new version is compatible with the repo.
Generate the wheels and verify that they work reading protos in a python
repo.
## Checklist:
- [x] I have self-reviewed this change.
- [x] I have tested this change.
- [ ] This change is covered by tests that are already landed, or in
this PR.